requirements.txt
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
flask-cors==4.0.0 